pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infrasys"
version = "0.2.0"
description = ''
readme = "README.md"
requires-python = ">=3.10, <3.13"
license = "BSD-3-Clause"
keywords = []
authors = [
    { name = "Aadil Latif", email = "aadil.latif@nrel.gov" },
    { name = "Daniel Thom", email = "daniel.thom@nrel.gov" },
    { name = "Kapil Duwadi", email = "kapil.duwadi@nrel.gov" },
    { name = "Pedro Andres Sanchez Perez", email = "pedroandres.sanchezperez@nrel.gov" },
    { name = "Tarek Elgindy", email = "tarek.elgindy@nrel.gov" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    # This is what we want, but are currently be limited by pyarrow.
    # Leave unbounded until pyarrow upgrades numpy.
    #"numpy >= 2, < 3",
    "numpy",
    "pyarrow~=15.0.2",
    "pint~=0.23",
    "loguru~=0.7.2",
    "rich~=13.7.1",
    "pydantic >= 2.7, < 3",
]
[project.optional-dependencies]
dev = [
    "furo",
    "mypy >=1.13, < 2",
    "myst_parser",
    "pandas-stubs",
    "pre-commit",
    "pyarrow-stubs",
    "pytest",
    "pytest-cov",
    "ruff",
    "sphinx",
    "sphinx-click",
    "autodoc_pydantic~=2.0",
    "sphinx-copybutton",
    "sphinx-tabs~=3.4",
]

[project.urls]
Documentation = "https://github.com/NREL/infrasys#readme"
Issues = "https://github.com/NREL/infrasys/issues"
Source = "https://github.com/NREL/infrasys"

[tool.mypy]
check_untyped_defs = true
files = [
  "src",
  "tests",
]

[tool.pytest.ini_options]
pythonpath = "src"
minversion = "6.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "_build",
    "build",
    "dist",
    "env",
    "venv",
]

line-length = 99
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "C901",   # McCabe complexity
    "E4",     # Subset of pycodestyle (E)
    "E7",
    "E9",
    "EM",     # string formatting in an exception message
    "F",      # Pyflakes
    "W",      # pycodestyle warnings
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]
